Update BellmanFord_Pyspark.py
Currently, the size of SparkGraph.graph_rdd is 2E where E is the edge vector until I can figure out how to 1.) get nodes that only appear as 'u' values to be included in distance calculations, and 2.) get a  factor of -1 applied to the weight when the ordered pair is inverted. Also, the current iteration does not discriminate on node type  (movie or a user); it calculates the distance for all of them. Since only movie node distances are of interest - and the number of users is typically orders of magnitude larger than that of movies - this should be avoided. The current workaround is to filter out user nodes post hoc.